+++ app/Models/Task.php
@@ -3,9 +3,12 @@
 namespace App\Models;
 
 use Illuminate\Database\Eloquent\Model;
+use Spatie\Activitylog\Traits\LogsActivity;
 
 class Task extends Model
 {
+    use LogsActivity;
+
     /**
      * The attributes that are mass assignable.
      *